backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "vetchat-ai-backend"
version = "0.1.0"
description = "VetChat AI — veterinary advising chat backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "pydantic-settings==2.7.0",
    "python-dotenv==1.0.1",
    "httpx==0.28.1",
    "anthropic>=0.40.0",
    "chromadb>=0.6.0",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.1",
]
